Feature/65

+++ graphql/resolvers/mutations/link.mutation.js
@@ -1,4 +1,5 @@
 const check = require('../../../helpers/check');
+const { getUserFromLink } = require('../../../helpers/get-user');
 
 module.exports = {
   createLink: async (
@@ -8,16 +9,18 @@ module.exports = {
   ) => {
     try {
       check.Auth(authUser);
-      return await db.links.create(
-        {
-          title,
-          url,
-          tags
-        },
-        {
-          include: [{ model: db.tags }]
-        }
-      );
+      if (authUser.sub === user.id) {
+        return await db.links.create(
+          {
+            title,
+            url,
+            tags
+          },
+          {
+            include: [{ model: db.tags }]
+          }
+        );
+      }
     } catch (error) {
       throw new Error(error);
     }
@@ -25,10 +28,13 @@ module.exports = {
   updateLink: async (_, { data: { id, title, url } }, { db, authUser }) => {
     try {
       check.Auth(authUser);
-      return await db.links.update(
-        { title: title, url: url },
-        { where: { id: id } }
-      );
+      const user = await getUserFromLink(id);
+      if (authUser.sub === user.id) {
+        return await db.links.update(
+          { title: title, url: url },
+          { where: { id: id } }
+        );
+      }
     } catch (error) {
       throw new Error(error);
     }
@@ -36,9 +42,12 @@ module.exports = {
   deleteLink: async (_, { data: { id } }, { db, authUser }) => {
     try {
       check.Auth(authUser);
-      return await db.links.destroy({ where: { id: id } });
+      const user = await getUserFromLink(id);
+      if (authUser.sub === user.id) {
+        return await db.links.destroy({ where: { id: id } });
+      }
     } catch (error) {
-      throw new error();
+      throw new Error(error);
     }
   }
 };
